Refactor reconstruction context and video export
