+++ DataAccessLayer/IReconstructionRepository.cs
@@ -6,7 +6,7 @@ namespace DataAccessLayer
 {
     public interface IReconstructionRepository
     {
-        void SaveReconstruction(List<ReconstructionResult> frames, string name, EITReconstructionParameters parameters);
+        void SaveReconstruction(List<ReconstructionResult> frames, string name, ReconstructionRuntimeContext parameters);
         IEnumerable<ReconstructionInfo> GetReconstructions();
         List<ReconstructionResult> LoadReconstruction(string filePath);
     }
